hash-parser: move `struct` and `enum` definition parsing into `definitions.rs`
