+++ compiler/hash-parser/src/parser/statement.rs
@@ -3,11 +3,8 @@
 //!
 //! All rights reserved 2022 (c) The Hash Language authors
 
-use hash_alloc::row;
 use hash_ast::ast::*;
-use hash_token::{delimiter::Delimiter, keyword::Keyword, Token, TokenKind, TokenKindVector};
-
-use crate::parser::error::TyArgumentKind;
+use hash_token::{keyword::Keyword, Token, TokenKind};
 
 use super::{error::AstGenErrorKind, AstGen, AstGenResult};
 
@@ -199,188 +196,5 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
             bound,
             trait_type,
         })
-    }
-
-    /// AST representation of a struct which includes the name of the struct with a
-    /// ForAll to specify any bounds or and generic arguments to the struct, with
-    /// zero or more struct fields. An example for a struct would be:
-    ///
-    /// struct Name <T,Q> where eq<T> { ... };
-    ///        ^^^^    ^──────^^─┬──^   ^^^
-    /// Name of struct        For all  fields
-    pub fn parse_struct_defn(
-        &self,
-        name: AstNode<'c, Name>,
-        bound: Option<AstNode<'c, Bound<'c>>>,
-    ) -> AstGenResult<'c, StructDef<'c>> {
-        debug_assert!(self
-            .current_token()
-            .has_kind(TokenKind::Keyword(Keyword::Struct)));
-
-        let entries = match self.peek() {
-            Some(token) if token.is_paren_tree() => self.parse_struct_def_entries()?,
-            token => self.error(
-                AstGenErrorKind::TyArgument(TyArgumentKind::Struct),
-                None,
-                token.map(|tok| tok.kind),
-            )?,
-        };
-
-        Ok(StructDef {
-            name,
-            bound,
-            entries,
-        })
-    }
-
-    /// Parse an enum definition. AST representation for an enum, An enum is constructed by
-    /// a the keyword 'enum' followed by an identifier name, a for-all declaration,
-    /// followed by some enum fields. An enumeration can be made of zero or more enum fields.
-    /// For example, a declaration of an enum would be:
-    ///
-    /// Name := <T, Q> where Eq<T> => enum(...);
-    /// ^^^^       ^──────^^─┬──^          ^^^
-    /// Name of enum      For all         fields
-    ///
-    pub fn parse_enum_defn(
-        &self,
-        name: AstNode<'c, Name>,
-        bound: Option<AstNode<'c, Bound<'c>>>,
-    ) -> AstGenResult<'c, EnumDef<'c>> {
-        debug_assert!(self
-            .current_token()
-            .has_kind(TokenKind::Keyword(Keyword::Enum)));
-
-        // now parse the optional type bound and the enum definition entries,
-        // if a type bound is specified, then the definition of the struct should
-        // be followed by an arrow ('=>')...
-        let entries = match self.peek() {
-            Some(token) if token.is_paren_tree() => self.parse_enum_def_entries()?,
-            token => self.error(
-                AstGenErrorKind::TyArgument(TyArgumentKind::Enum),
-                None,
-                token.map(|tok| tok.kind),
-            )?,
-        };
-
-        Ok(EnumDef {
-            name,
-            bound,
-            entries,
-        })
-    }
-
-    /// Function used to pass a comma separated [EnumDefEntry] which is proceeded
-    /// within the beginning of an [EnumDef].
-    pub fn parse_enum_def_entries(&self) -> AstGenResult<'c, AstNodes<'c, EnumDefEntry<'c>>> {
-        match self.peek() {
-            Some(Token {
-                kind: TokenKind::Tree(Delimiter::Paren, tree_index),
-                span,
-            }) => {
-                self.skip_token();
-                let tree = self.token_trees.get(*tree_index).unwrap();
-                let gen = self.from_stream(tree, *span);
-
-                gen.parse_separated_fn(
-                    || gen.parse_enum_def_entry(),
-                    || gen.parse_token_atom(TokenKind::Comma),
-                )
-            }
-            Some(token) => self.error(
-                AstGenErrorKind::Expected,
-                Some(TokenKindVector::singleton(
-                    &self.wall,
-                    TokenKind::Delimiter(Delimiter::Brace, false),
-                )),
-                Some(token.kind),
-            ),
-            None => self.unexpected_eof(),
-        }
-    }
-
-    /// Parse an [EnumDefEntry].
-    pub fn parse_enum_def_entry(&self) -> AstGenResult<'c, AstNode<'c, EnumDefEntry<'c>>> {
-        let name = self.parse_name()?;
-        let name_location = name.location();
-
-        let mut args = AstNodes::empty();
-
-        if let Some(Token {
-            kind: TokenKind::Tree(Delimiter::Paren, tree_index),
-            span,
-        }) = self.peek()
-        {
-            self.skip_token();
-            args.span = Some(*span);
-            let tree = self.token_trees.get(*tree_index).unwrap();
-
-            let gen = self.from_stream(tree, *span);
-            while gen.has_token() {
-                let ty = gen.parse_type()?;
-                args.nodes.push(ty, &self.wall);
-
-                if gen.has_token() {
-                    gen.parse_token_atom(TokenKind::Comma)?;
-                }
-            }
-        }
-
-        Ok(self.node_with_joined_span(EnumDefEntry { name, args }, &name_location))
-    }
-
-    /// Function used to pass a comma separated [StructDefEntry] which is proceeded
-    /// within the beginning of an [StructDef].
-    pub fn parse_struct_def_entries(&self) -> AstGenResult<'c, AstNodes<'c, StructDefEntry<'c>>> {
-        match self.peek() {
-            Some(Token {
-                kind: TokenKind::Tree(Delimiter::Paren, tree_index),
-                span,
-            }) => {
-                self.skip_token();
-
-                let tree = self.token_trees.get(*tree_index).unwrap();
-                let gen = self.from_stream(tree, *span);
-
-                gen.parse_separated_fn(
-                    || gen.parse_struct_def_entry(),
-                    || gen.parse_token_atom(TokenKind::Comma),
-                )
-            }
-            Some(token) => {
-                let atom = token.kind;
-                let expected = TokenKindVector::from_row(
-                    row![&self.wall; TokenKind::Delimiter(Delimiter::Brace, true)],
-                );
-
-                self.error(AstGenErrorKind::Expected, Some(expected), Some(atom))?
-            }
-            None => self.unexpected_eof(),
-        }
-    }
-
-    /// Parse a [StructDefEntry].
-    pub fn parse_struct_def_entry(&self) -> AstGenResult<'c, AstNode<'c, StructDefEntry<'c>>> {
-        let start = self.current_location();
-        let name = self.parse_name()?;
-
-        let ty = match self.peek() {
-            Some(token) if token.has_kind(TokenKind::Colon) => {
-                self.skip_token();
-                Some(self.parse_type()?)
-            }
-            _ => None,
-        };
-
-        let default = match self.peek() {
-            Some(token) if token.has_kind(TokenKind::Eq) => {
-                self.skip_token();
-
-                Some(self.parse_expression_with_precedence(0)?)
-            }
-            _ => None,
-        };
-
-        Ok(self.node_with_joined_span(StructDefEntry { name, ty, default }, &start))
     }
 }
